fixed checkstyle issues in adapters for search by airline id
