+++ src/java/adapters/search_by_airlineid/SearchByAirlineIDController.java
@@ -1,4 +1,4 @@
-package adapters.SearchByAirlineID;
+package adapters.search_by_airlineid;
 
 import use_case.SearchByAirlineID.SearchByAirlineIDInputBoundary;
 import use_case.SearchByAirlineID.SearchByAirlineIDInputData;
@@ -17,13 +17,15 @@ public class SearchByAirlineIDController {
     /**
      * Handles the user request to search for flights by an airline's IATA code.
      * @param airlineIataCode the airline IATA code entered by the user.
+     * @throws IllegalArgumentException when the Airport IATA code is empty
      */
     public void execute(String airlineIataCode) {
+        // Checks if the airport IATA Code is null and if it is, throws a IllegalArgumentException
         if (airlineIataCode == null || airlineIataCode.trim().isEmpty()) {
             throw new IllegalArgumentException("Airline IATA code cannot be empty.");
         }
 
-        // Create input data and call the input boundary
+        // Creates input data and then excecutes the Interactor with the input boundary
         SearchByAirlineIDInputData inputData = new SearchByAirlineIDInputData(airlineIataCode.trim());
         searchByAirlineIDInteractor.execute(inputData);
     }
